requirements.txt
# ------------- Scraping Engine -------------
scrapy==2.11.1
scrapy-playwright==0.0.33
playwright==1.42.0

# ------------- Data Validation -------------
pydantic

# ------------- Database & ORM -------------
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.11
alembic>=1.13.1

# ------------- Orchestration -------------
celery==5.3.6
redis==5.0.3

# ------------- API Layer -------------
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
